+++ tests/Unit/EnumTest.php
@@ -20,6 +20,8 @@ class EnumTest extends TestCase
             'APPLE',
             'BANANA',
             'CHERRY',
+            'EGGPLANT',
+            'AUBERGINE',
         ], Fruit::names());
 
         $this->assertEquals([
@@ -36,6 +38,8 @@ class EnumTest extends TestCase
             Fruit::APPLE,
             Fruit::BANANA,
             Fruit::CHERRY,
+            Fruit::EGGPLANT,
+            Fruit::AUBERGINE,
         ], Fruit::keys());
 
         $this->assertEquals([
@@ -55,12 +59,14 @@ class EnumTest extends TestCase
 
     public function test_get_values()
     {
-        // When map() is not defined, should return an array of null
+        // Number does not provide a map() method and therefore all keys are
+        // by default mapped to a value of null
         $this->assertEquals([
             null,
             null,
             null,
-        ], Fruit::values());
+            null,
+        ], Number::values());
 
         // When map() is defined, should return all of the mapped values
         $this->assertEquals([
@@ -74,7 +80,7 @@ class EnumTest extends TestCase
     public function test_can_get_value_for_key()
     {
         // Not mapped
-        $this->assertEquals(null, Fruit::valueForKey(Fruit::APPLE));
+        $this->assertEquals(null, Number::valueForKey(Number::TWENTY_FOUR));
 
         // Mapped
         $this->assertEquals('Kitty-cat', Animal::valueForKey(Animal::CAT));
@@ -128,9 +134,8 @@ class EnumTest extends TestCase
 
     public function test_can_get_value_from_instance()
     {
-        $this->assertEquals(null, Fruit::APPLE()->value());
-        $this->assertEquals(null, Fruit::BANANA()->value());
-
+        $this->assertEquals('Apple', Fruit::APPLE()->value());
+        $this->assertEquals(null, Number::TWENTY_FOUR()->value());
         $this->assertEquals('Kitty-cat', Animal::CAT()->value());
         $this->assertEquals(null, Animal::HORSE()->value());
         $this->assertEquals(['you', 'filthy', 'animal'], Animal::PIGEON()->value());
@@ -177,27 +182,22 @@ class EnumTest extends TestCase
         $this->assertEquals(Fruit::APPLE()->name(), (string)Fruit::APPLE());
     }
 
-    public function test_flipable_trait_flips_map()
+    public function test_get_key_by_value()
     {
-        $this->assertEquals([
-            'Corona' => Bevs::BREW,
-            'Red Wine' => Bevs::RED_WINE,
-            'White Wine' => Bevs::WHITE_WINE,
-            'Bundaberg' => Bevs::RUM,
-            'Jack Daniels' => Bevs::BOURBON,
-        ], Bevs::flip());
+        $this->assertEquals(Bevs::BREW, Bevs::keyForValue('Corona'));
+        $this->assertEquals(Bevs::RUM, Bevs::keyForValue('Bundaberg'));
     }
 
-    public function test_flipable_trait_gets_constant_by_value()
-    {
-        $this->assertEquals(Bevs::BREW, Bevs::fromValue('Corona'));
-        $this->assertEquals(Bevs::RUM, Bevs::fromValue('Bundaberg'));
-    }
-
-    public function test_flipable_trait_throws_exception_with_invalid_value()
+    public function test_get_key_by_invalid_value_throws_exception()
     {
         $this->expectException(InvalidValueException::class);
-        $this->assertEquals(Bevs::BREW, Bevs::fromValue('Water'));
+        $this->assertEquals(Bevs::BREW, Bevs::keyForValue('Water'));
+    }
+
+    public function test_get_key_by_duplicate_value_returns_first()
+    {
+        // Fruit::EGGPLANT and Fruit::AUBERGINE are both mapped to 'Eggplant'
+        $this->assertEquals(Fruit::EGGPLANT, Fruit::keyForValue('Eggplant'));
     }
 
     public function test_get_instance_via_name()